+++ src/defaultRanges.js
@@ -101,6 +101,7 @@ export const defaultInputRanges = [
     },
     getCurrentValue(range) {
       if (!isSameDay(range.endDate, defineds.endOfToday)) return '-';
+      if (!range.startDate) return '∞';
       return differenceInCalendarDays(defineds.endOfToday, range.startDate);
     },
   },
@@ -115,6 +116,7 @@ export const defaultInputRanges = [
     },
     getCurrentValue(range) {
       if (!isSameDay(range.startDate, defineds.startOfToday)) return '-';
+      if (!range.endDate) return '∞';
       return differenceInCalendarDays(range.endDate, defineds.startOfToday);
     },
   },